blog-service/app/requirements.txt
Django>=4.2.0,<5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-redis==5.4.0
django-filter==23.3
python-slugify==8.0.1
gunicorn==21.2.0